Create README.md
The Central Weather Administration (CWA) of Taiwan provides real-time weather conditions, 3-day and 7-day weather forecasts, and satellite weather images of Taiwan. We develop a MCP server to access the CWA services. A large language model (LLM) can use the tool in response to user query about the weather of and near Taiwan.